shenzhen-survival-sim/requirements.txt
# 深圳生存模拟 - 后端依赖
# 安装: pip install -r requirements.txt

fastapi>=0.100.0
uvicorn>=0.22.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
